+++ packages/yaml/src/Conferer/Source/Yaml.hs
@@ -14,17 +14,21 @@ import qualified Conferer.Source.Aeson as JSON
 import Conferer.Source.Files
 import Conferer.Source
 
--- | Create a 'SourceCreator' from a yaml file that we get from the env 
+-- | Create a 'SourceCreator' from a yaml file that we get from the env
 -- with the same logic as 'JSON.JSONSource'
 fromConfig :: Key -> SourceCreator
 fromConfig key config = do
   filePath <- getFilePathFromEnv key "yaml" config
-  fromFilePath filePath
+  fromFilePath' filePath
 
+-- | Create a 'SourceCreator' by reading the provided path as json
+fromFilePath :: FilePath -> SourceCreator
+fromFilePath filePath _config =
+  fromFilePath' filePath
 
 -- | Create a 'Source' by reading the provided path as json 
-fromFilePath :: FilePath -> IO Source
-fromFilePath filePath = do
+fromFilePath' :: FilePath -> IO Source
+fromFilePath' filePath = do
   configAsJson <- decodeFileEither filePath
   case configAsJson of
     Right jsonConfig -> return $ JSON.fromValue jsonConfig